fix: improve structure for rendering
