+++ package.json
@@ -18,12 +18,15 @@
     "ios/",
     "touchables/**/*.js",
     "Directions.js",
+    "Directions.web.js",
     "DrawerLayout.js",
     "GestureHandler.js",
     "GestureHandlerButton.js",
     "GestureHandlerButton.web.js",
     "GestureHandlerModule.js",
     "GestureHandlerModule.web.js",
+    "GestureHandlerRenderer.js",
+    "GestureHandlerRenderer.web.js",
     "gestureHandlerRootHOC.android.js",
     "gestureHandlerRootHOC.ios.js",
     "gestureHandlerRootHOC.js",
